Moved createTrackForm function to the addTrackForm file to clean up file structure. #3
