+++ client/utils/addTrackForm.js
@@ -1,8 +1,23 @@
-const createTrackForm = require('./createTrackForm.js')
+const createElement = require('./createElement')
 
 function addTrackForm(currentTrack) {
   const $trackFormContainer = document.getElementById('track-form-container')
   $trackFormContainer.appendChild(createTrackForm(currentTrack))
+}
+
+function createTrackForm(currentTrack) {
+  const trackFormFields = ['num', 'name', 'start', 'end']
+  const $trackForm = createElement('tr', {class: 'track-form-' + currentTrack}, '', [])
+  for (let i = 0; i < trackFormFields.length; i++) {
+    const $tableCell = createElement('td', {}, '', [])
+    const $trackFormField = createElement('input', {
+      id: 'track-' + trackFormFields[i] + '-' + currentTrack,
+      name: 'track-' + trackFormFields[i] + '-' + currentTrack
+    }, '', [])
+    $tableCell.appendChild($trackFormField)
+    $trackForm.appendChild($tableCell)
+  }
+  return $trackForm
 }
 
 module.exports = addTrackForm
