--- client/utils/createTrackForm.js
@@ -1,18 +1,0 @@
-const createElement = require('./createElement')
-
-function createTrackForm(currentTrack) {
-  const trackFormFields = ['num', 'name', 'start', 'end']
-  const $trackForm = createElement('tr', {class: 'track-form-' + currentTrack}, '', [])
-  for (let i = 0; i < trackFormFields.length; i++) {
-    const $tableCell = createElement('td', {}, '', [])
-    const $trackFormField = createElement('input', {
-      id: 'track-' + trackFormFields[i] + '-' + currentTrack,
-      name: 'track-' + trackFormFields[i] + '-' + currentTrack
-    }, '', [])
-    $tableCell.appendChild($trackFormField)
-    $trackForm.appendChild($tableCell)
-  }
-  return $trackForm
-}
-
-module.exports = createTrackForm
